+++ packages/amplify-ui-components/src/components/amplify-greetings/amplify-greetings.stories.tsx
@@ -1,5 +1,0 @@
-export default {
-	title: 'amplify-greetings',
-};
-
-export const defaults = () => `<amplify-greetings />`;