+++ packages/analytics/src/Providers/AmazonPersonalizeHelper/SessionInfoManager.ts
@@ -1,22 +1,11 @@
-/*
- * Copyright 2019-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
- * the License. A copy of the License is located at
- *
- *     http://aws.amazon.com/apache2.0/
- *
- * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
- * and limitations under the License.
- */
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
 import { SessionInfo } from './DataType';
 import isEmpty from 'lodash/isEmpty';
 import isEqual from 'lodash/isEqual';
 import { v1 as uuid } from 'uuid';
-import { ConsoleLogger as Logger, JS } from '@aws-amplify/core';
-
-import Cache from '@aws-amplify/cache';
+import { ConsoleLogger as Logger, browserOrNode } from '@aws-amplify/core';
+import { Cache } from '@aws-amplify/cache';
 
 const PERSONALIZE_CACHE = '_awsct';
 const PERSONALIZE_CACHE_USERID = '_awsct_uid';
@@ -34,7 +23,7 @@ export class SessionInfoManager {
 	private _timerKey;
 
 	constructor(prefixKey = '') {
-		this._isBrowser = JS.browserOrNode().isBrowser;
+		this._isBrowser = browserOrNode().isBrowser;
 		this._timerKey = uuid().substr(0, 15);
 		this._refreshTimer();
 	}