+++ packages/analytics/src/trackers/EventTracker.ts
@@ -1,19 +1,9 @@
-/*
- * Copyright 2017-2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
- * the License. A copy of the License is located at
- *
- *     http://aws.amazon.com/apache2.0/
- *
- * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
- * and limitations under the License.
- */
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
 
 import { delegate } from '../vendor/dom-utils';
 import { EventTrackOpts } from '../types';
-import { ConsoleLogger as Logger, JS } from '@aws-amplify/core';
+import { ConsoleLogger as Logger, browserOrNode } from '@aws-amplify/core';
 
 const logger = new Logger('EventTracker');
 
@@ -30,7 +20,7 @@ export class EventTracker {
 	private _delegates;
 
 	constructor(tracker, opts) {
-		if (!JS.browserOrNode().isBrowser || !window.addEventListener) {
+		if (!browserOrNode().isBrowser || !window.addEventListener) {
 			logger.debug('not in the supported web environment');
 			return;
 		}
@@ -123,8 +113,3 @@ export class EventTracker {
 		});
 	}
 }
-
-/**
- * @deprecated use named import
- */
-export default EventTracker;