+++ packages/analytics/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.27](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.26...@aws-amplify/analytics@5.2.27) (2022-10-27)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.26](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.25...@aws-amplify/analytics@5.2.26) (2022-10-26)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.25](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.24...@aws-amplify/analytics@5.2.25) (2022-10-25)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.24](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.23...@aws-amplify/analytics@5.2.24) (2022-10-14)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
+## [5.2.23](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.22...@aws-amplify/analytics@5.2.23) (2022-10-14)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
 ## [5.2.22](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.20...@aws-amplify/analytics@5.2.22) (2022-09-30)
 
 **Note:** Version bump only for package @aws-amplify/analytics