+++ packages/cache/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.62](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@4.0.61...@aws-amplify/cache@4.0.62) (2022-10-27)
+
+**Note:** Version bump only for package @aws-amplify/cache
+
+
+
+
+
+## [4.0.61](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@4.0.60...@aws-amplify/cache@4.0.61) (2022-10-26)
+
+**Note:** Version bump only for package @aws-amplify/cache
+
+
+
+
+
+## [4.0.60](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@4.0.59...@aws-amplify/cache@4.0.60) (2022-10-25)
+
+**Note:** Version bump only for package @aws-amplify/cache
+
+
+
+
+
+## [4.0.59](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@4.0.58...@aws-amplify/cache@4.0.59) (2022-10-14)
+
+**Note:** Version bump only for package @aws-amplify/cache
+
+
+
+
+
+## [4.0.58](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@4.0.57...@aws-amplify/cache@4.0.58) (2022-10-14)
+
+**Note:** Version bump only for package @aws-amplify/cache
+
+
+
+
+
 ## [4.0.57](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/cache@4.0.55...@aws-amplify/cache@4.0.57) (2022-09-30)
 
 **Note:** Version bump only for package @aws-amplify/cache