+++ packages/geo/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.23](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@1.3.22...@aws-amplify/geo@1.3.23) (2022-10-27)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+
+
+
+
+## [1.3.22](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@1.3.21...@aws-amplify/geo@1.3.22) (2022-10-26)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+
+
+
+
+## [1.3.21](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@1.3.20...@aws-amplify/geo@1.3.21) (2022-10-25)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+
+
+
+
+## [1.3.20](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@1.3.19...@aws-amplify/geo@1.3.20) (2022-10-14)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+
+
+
+
+## [1.3.19](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@1.3.18...@aws-amplify/geo@1.3.19) (2022-10-14)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
+
+
+
+
 ## [1.3.18](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@1.3.16...@aws-amplify/geo@1.3.18) (2022-09-30)
 
 **Note:** Version bump only for package @aws-amplify/geo