+++ packages/interactions/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.8](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/interactions@4.1.7...@aws-amplify/interactions@4.1.8) (2022-10-27)
+
+**Note:** Version bump only for package @aws-amplify/interactions
+
+
+
+
+
+## [4.1.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/interactions@4.1.6...@aws-amplify/interactions@4.1.7) (2022-10-26)
+
+**Note:** Version bump only for package @aws-amplify/interactions
+
+
+
+
+
+## [4.1.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/interactions@4.1.5...@aws-amplify/interactions@4.1.6) (2022-10-25)
+
+**Note:** Version bump only for package @aws-amplify/interactions
+
+
+
+
+
+## [4.1.5](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/interactions@4.1.4...@aws-amplify/interactions@4.1.5) (2022-10-14)
+
+**Note:** Version bump only for package @aws-amplify/interactions
+
+
+
+
+
+## [4.1.4](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/interactions@4.1.3...@aws-amplify/interactions@4.1.4) (2022-10-14)
+
+**Note:** Version bump only for package @aws-amplify/interactions
+
+
+
+
+
 ## [4.1.3](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/interactions@4.1.1...@aws-amplify/interactions@4.1.3) (2022-09-30)